GI.Gtk.Declarative

Description

The declarative layer on top of GTK+ lets you describe your user interface as a declarative hierarchy of objects, using data structures and pure functions. You can leverage the declarative event handling to build reusable widgets. The Patch typeclass, and the instances provided by this library, performs minimal updates to GTK+ widgets using the underlying imperative operations, so that your rendering can always be a pure function your state to a Widget.

Documentation

data EventHandler gtkEventHandler widget (purity :: Purity) event where

Encodes the user event handler in such a way that we can have a Functor instance for arity-polymorphic event handlers.

Constructors

PureEventHandler :: EventHandlerReturn Identity ret e -> EventHandler (IO ret) w Pure e

ImpureEventHandler :: (w -> EventHandlerReturn IO ret e) -> EventHandler (IO ret) w Impure e

EventHandlerFunction :: (a -> EventHandler b w p e) -> EventHandler (a -> b) w p e

class (Typeable decl, Typeable (AttrState decl)) => CustomAttribute widget decl where

Defines types that can be used as declarative custom attributes on the given widget type.

Minimal complete definition

attrCreate, attrPatch

Associated Types

data AttrState decl

The runtime state of this attribute. This is preserved between calls to attrCreateattrPatchattrDestroy.

Methods

attrCreate :: ComponentContext -> widget -> decl event -> IO (AttrState decl)

Called when this attribute is first attached to a widget in the tree

attrPatch :: ComponentContext -> widget -> AttrState decl -> decl event1 -> decl event2 -> IO (AttrState decl)

Called when the widget tree is being patched.

attrDestroy :: ComponentContext -> widget -> AttrState decl -> decl event -> IO ()

Called when the associated widget is removed from the widget tree.

attrSubscribe :: ComponentContext -> widget -> AttrState decl -> decl event -> (event -> IO ()) -> IO Subscription

Attach event handlers to this attribute.

data CustomAttributeKey where

A key for a custom attribute. The actual type of the key is hidden: we only care that it is typeable and hashable.

Constructors

CustomAttributeKey :: (Typeable key, Eq key, Hashable key) => key -> CustomAttributeKey

type CollectedCustomAttributeStates widget = HashMap CustomAttributeKey [CustomAttributeState widget]

This is the state that is maintained for the custom attributes on a particular widget. It is grouped by key for the convenience of the diffing algorithm.

data CustomAttributeState widget where

The runtime state for a custom attribute. The actual type of the attribute is hidden here: only the widget type is exposed.

Constructors

CustomAttributeState :: CustomAttribute widget decl => AttrState decl -> CustomAttributeState widget

data CustomAttributeDecl widget event where

The declarative form of a custom attribute. The actual type of of the attribute is hidden here: only the widget and event types are exposed.

Constructors

CustomAttributeDecl :: CustomAttribute widget decl => CustomAttributeKey -> decl event -> CustomAttributeDecl widget event

createCustomAttributes :: ComponentContext -> widget -> Vector (CustomAttributeDecl widget event) -> IO (CollectedCustomAttributeStates widget)

Runs the create action for each custom attribute.

patchCustomAttributes :: forall widget e1 e2. ComponentContext -> widget -> CollectedCustomAttributeStates widget -> Vector (CustomAttributeDecl widget e1) -> Vector (CustomAttributeDecl widget e2) -> IO (CollectedCustomAttributeStates widget)

Patches custom attribute state so it matches the latest declarative attributes.

destroyCustomAttributes :: ComponentContext -> widget -> CollectedCustomAttributeStates widget -> Vector (CustomAttributeDecl widget event) -> IO ()

Runs the destroy action for the given custom attributes.

subscribeCustomAttributes :: ComponentContext -> widget -> CollectedCustomAttributeStates widget -> Vector (CustomAttributeDecl widget event) -> (event -> IO ()) -> IO Subscription

Attaches event listeners to already-created custom attributes.

type ClassSet = HashSet Text

A set of CSS classes.

data Attribute widget event where

The attribute GADT represent a supported attribute for a declarative widget. This extends the regular notion of GTK+ attributes to also include event handling and CSS classes.

Constructors

(:=) :: (AttrOpAllowed 'AttrConstruct info widget, AttrOpAllowed 'AttrSet info widget, AttrGetC info widget attr getValue, AttrSetTypeConstraint info setValue, KnownSymbol attr, Typeable attr, Eq setValue, Typeable setValue) => AttrLabelProxy (attr :: Symbol) -> setValue -> Attribute widget event
An attribute/value mapping for a declarative widget. The AttrLabelProxy is parameterized by attr, which represents the GTK-defined attribute name. The underlying GI object needs to support the construct, get, and set operations for the given attribute.

Classes :: IsWidget widget => ClassSet -> Attribute widget event
Defines a set of CSS classes for the underlying widget's style context. Use the classes function instead of this constructor directly.

OnSignalPure :: (GObject widget, SignalInfo info, gtkCallback ~ HaskellCallbackType info, ToGtkCallback gtkCallback Pure) => SignalProxy widget info -> EventHandler gtkCallback widget Pure event -> Attribute widget event
Emit events using a pure event handler. Use the on function, instead of this constructor directly.

OnSignalImpure :: (GObject widget, SignalInfo info, gtkCallback ~ HaskellCallbackType info, ToGtkCallback gtkCallback Impure) => SignalProxy widget info -> EventHandler gtkCallback widget Impure event -> Attribute widget event
Emit events using a pure event handler. Use the on function, instead of this constructor directly.

Custom :: CustomAttributeDecl widget event -> Attribute widget event

classes :: IsWidget widget => [Text] -> Attribute widget event

Define the CSS classes for the underlying widget's style context. For these classes to have any effect, this requires a CssProvider with CSS files loaded, to be added to the GDK screen. You probably want to do this in your entry point when setting up GTK.

on :: (GObject widget, SignalInfo info, gtkCallback ~ HaskellCallbackType info, ToGtkCallback gtkCallback Pure, ToEventHandler gtkCallback widget Pure, userEventHandler ~ UserEventHandler gtkCallback widget Pure event) => SignalProxy widget info -> userEventHandler -> Attribute widget event

Emit events, using a pure event handler, by subcribing to the specified signal.

onM :: (GObject widget, SignalInfo info, gtkCallback ~ HaskellCallbackType info, ToGtkCallback gtkCallback Impure, ToEventHandler gtkCallback widget Impure, userEventHandler ~ UserEventHandler gtkCallback widget Impure event) => SignalProxy widget info -> userEventHandler -> Attribute widget event

Emit events, using an impure event handler receiving the widget and returning an IO action of event, by subcribing to the specified signal.

customAttribute :: (Typeable key, Eq key, Hashable key, CustomAttribute widget decl) => key -> decl event -> Attribute widget event

Create a custom attribute from its declarative form

filterToCustom :: Vector (Attribute widget event) -> Vector (CustomAttributeDecl widget event)

Extract the custom attributes from a list of attributes.

class Patchable widget where

A patchable widget is one that can create an underlying GTK widget, or calculate a Patch to be applied to an existing GTK widget that was previously created.

Methods

create :: ComponentContext -> widget e -> IO SomeState

Given a declarative widget that is Patchable, return an IO action that can create a new corresponding Widget. The created widget should be use in corresponding patch modifications, until it is replaced.

patch :: ComponentContext -> SomeState -> widget e1 -> widget e2 -> Patch

Given two declarative widgets of the same widget type (but not necessarily of the same event types,) calculate a Patch.

destroy :: ComponentContext -> SomeState -> widget e -> IO ()

Given a previously created declarative widget, destroy it.

data Patch

A possible action to take on an existing Widget, decided by the patch method when comparing declarative widgets.

Constructors

Modify (IO SomeState)
An IO action to apply to a Widget to make it reflect an updated declarative widget. The action to apply is calculated from the difference between the old and the new declarative widget.

Replace (IO SomeState)
Replace the current Widget by the widget returned by the IO action.

Keep
Do nothing, i.e. keep the Widget as it is.

class FromWidget widget target where

Convert a widget to a target type. This is deliberately unconstrained in it's types, and is used by smart constructors to implement return type polymorphism, so that a smart contructor can return either a Widget, or some specifically typed widget, depending on the context in which it's used.

Methods

fromWidget :: widget event -> target event

data Widget event where

A Widget value wraps a Patchable and EventSource widget, providing a constrained equivalent of a Dynamic value. It is used to support heterogeneous containers of widgets, and to support equality checks on different types of widgets when calculating patches.

Constructors

Widget :: (Typeable widget, Patchable widget, EventSource widget) => widget event -> Widget event

data SingleWidget widget event

Declarative version of a leaf widget, i.e. a widget without any children.

widget

Arguments

:: (Typeable widget, IsWidget widget, FromWidget (SingleWidget widget) target)

=> (ManagedPtr widget -> widget)
A widget constructor from the underlying gi-gtk library.

-> Vector (Attribute widget event)
List of Attributes.

-> target event
The target, whose type is decided by FromWidget.

Construct a leaf widget, i.e. one without any children.

data BoxChildProperties

Values used when packing child widgets into boxes.

Constructors

BoxChildProperties

Fields

expand :: Bool

fill :: Bool

padding :: Word32

data BoxChild event

Describes a child widget to be added with boxAppend to a Box.

Constructors

BoxChild

Fields

properties :: BoxChildProperties

child :: Widget event

defaultBoxChildProperties :: BoxChildProperties

Defaults for BoxChildProperties. Use these and override specific fields.

data Container widget children event

Declarative version of a container widget, i.e. a widget with zero or more child widgets. The type of children is parameterized, and differs across the supported container widgets, as some containers require specific types of child widgets. These type relations are decided by IsContainer, and instances can found in GI.Gtk.Declarative.Container.Patch.

container

Arguments

:: (Typeable widget, IsWidget widget, IsContainer widget, FromWidget (Container widget (Children child)) target, ToChildren widget parent child)

=> (ManagedPtr widget -> widget)
A container widget constructor from the underlying gi-gtk library.

-> Vector (Attribute widget event)
Attributes.

-> parent (child event)
The container's child widgets, in a MarkupOf builder.

-> target event
The target, whose type is decided by FromWidget.

Construct a container widget, i.e. a widget with zero or more children.

data PaneProperties

Values used when packing a pane into a Paned.

Constructors

PaneProperties

Fields

resize :: Bool

shrink :: Bool

data Pane event

Describes a pane to be packed with panePack1/panePack2 in a Paned.

defaultPaneProperties :: PaneProperties

Defaults for PaneProperties. Use these and override specific fields.

pane :: PaneProperties -> Widget event -> Pane event

Construct a pane to be packed with panePack1/panePack2 in a Paned.

paned :: Vector (Attribute Paned event) -> Pane event -> Pane event -> Widget event

Construct a Paned based on attributes and two child Panes.

data Page event

Describes a page to be added to a Notebook

page :: Text -> Widget event -> Page event

Create a page with a textual label and an arbitrary content widget.

pageWithTab :: Widget event -> Widget event -> Page event

Create a page with arbitrary widgets for both label and content.

notebook :: Vector (Attribute Notebook event) -> Vector (Page event) -> Widget event

Create a Notebook by combining multiple pages.

data Bin widget event

Declarative version of a bin widget, i.e. a widget with exactly one child.

bin

Arguments

:: (Typeable widget, IsContainer widget, IsBin widget, IsWidget widget, FromWidget (Bin widget) target)

=> (ManagedPtr widget -> widget)
A bin widget constructor from the underlying gi-gtk library.

-> Vector (Attribute widget event)
List of Attributes.

-> Widget event
The bin's child widget

-> target event
The target, whose type is decided by FromWidget.

Construct a bin widget, i.e. a widget with exactly one child.

data MenuItem event

A menu item widget used for Menu children.

menuItem :: (IsMenuItem item, Typeable item, IsContainer item, IsBin item, IsWidget item) => (ManagedPtr item -> item) -> Vector (Attribute item event) -> Widget event -> MenuItem event

Construct a single menu item for a Menu.

subMenu :: Text -> Vector (MenuItem event) -> MenuItem event

Construct a sub menu for a Menu, with a text label and the child menu items.