GI.Gtk.Declarative.Attributes.Custom

Description

Custom attributes that have arbitrary state and patching behaviour, yet can be attached to declarative widgets just like normal Gtk+ attributes.

Documentation

class (Typeable decl, Typeable (AttrState decl)) => CustomAttribute widget decl where

Defines types that can be used as declarative custom attributes on the given widget type.

Minimal complete definition

attrCreate, attrPatch

Associated Types

data AttrState decl

The runtime state of this attribute. This is preserved between calls to attrCreateattrPatchattrDestroy.

Methods

attrCreate :: ComponentContext -> widget -> decl event -> IO (AttrState decl)

Called when this attribute is first attached to a widget in the tree

attrPatch :: ComponentContext -> widget -> AttrState decl -> decl event1 -> decl event2 -> IO (AttrState decl)

Called when the widget tree is being patched.

attrDestroy :: ComponentContext -> widget -> AttrState decl -> decl event -> IO ()

Called when the associated widget is removed from the widget tree.

attrSubscribe :: ComponentContext -> widget -> AttrState decl -> decl event -> (event -> IO ()) -> IO Subscription

Attach event handlers to this attribute.

data CustomAttributeKey where

A key for a custom attribute. The actual type of the key is hidden: we only care that it is typeable and hashable.

Constructors

CustomAttributeKey :: (Typeable key, Eq key, Hashable key) => key -> CustomAttributeKey

data CustomAttributeDecl widget event where

The declarative form of a custom attribute. The actual type of of the attribute is hidden here: only the widget and event types are exposed.

Constructors

CustomAttributeDecl :: CustomAttribute widget decl => CustomAttributeKey -> decl event -> CustomAttributeDecl widget event

data CustomAttributeState widget where

The runtime state for a custom attribute. The actual type of the attribute is hidden here: only the widget type is exposed.

Constructors

CustomAttributeState :: CustomAttribute widget decl => AttrState decl -> CustomAttributeState widget

type CollectedCustomAttributeStates widget = HashMap CustomAttributeKey [CustomAttributeState widget]

This is the state that is maintained for the custom attributes on a particular widget. It is grouped by key for the convenience of the diffing algorithm.

createCustomAttributes :: ComponentContext -> widget -> Vector (CustomAttributeDecl widget event) -> IO (CollectedCustomAttributeStates widget)

Runs the create action for each custom attribute.

patchCustomAttributes :: forall widget e1 e2. ComponentContext -> widget -> CollectedCustomAttributeStates widget -> Vector (CustomAttributeDecl widget e1) -> Vector (CustomAttributeDecl widget e2) -> IO (CollectedCustomAttributeStates widget)

Patches custom attribute state so it matches the latest declarative attributes.

destroyCustomAttributes :: ComponentContext -> widget -> CollectedCustomAttributeStates widget -> Vector (CustomAttributeDecl widget event) -> IO ()

Runs the destroy action for the given custom attributes.

subscribeCustomAttributes :: ComponentContext -> widget -> CollectedCustomAttributeStates widget -> Vector (CustomAttributeDecl widget event) -> (event -> IO ()) -> IO Subscription

Attaches event listeners to already-created custom attributes.