GI.Gtk.Declarative.Attributes.Custom.Window

Documentation

window :: (Typeable key, Eq key, Hashable key) => key -> Bin Window event -> Attribute widget event

Construct a new declarative top-level window, with a lifecycle tied to the widget the attribute is attached to.

The key should uniquely identify this window amongst all windows attached to the same widget.

presentWindow :: (Eq a, Typeable a) => a -> Attribute Window event

Create an attribute for "presenting" (i.e. focusing) a window: when the value changes then the window will be presented.

data IconData

Constructors

IconDataBytes ByteString
Icon data in an image format supported by GdkPixbuf (e.g. PNG, JPG)

windowIcon :: IconData -> Attribute Window event

Set the icon that is used by one particular window.

setDefaultIcon :: IconData -> IO ()

Set the icon that is used for windows by default.