GI.Gtk.Declarative.Attributes.Internal.Conversions

Documentation

class ToGtkCallback gtkCallback purity where

Internal class for converting EventHandlers to gi-gtk callbacks.

Methods

toGtkCallback :: EventHandler gtkCallback widget purity event -> widget -> (event -> IO ()) -> gtkCallback

Converts an EventHandler, i.e. the internal encoding of a pure or an impure callback, back to a GTK+ callback. Impure callbacks will also receive a widget as the last argument.