GI.Gtk.Declarative.Component

Description

Declarative, composable, components with independent internal state and events.

Documentation

class Typeable c => Component (c :: * -> *) where

A declarative component that can be used as a widget via the component function. Types in this class are parameterized by the type of the events that the component emits externally (i.e. the type that its parent receives).

Minimal complete definition

createComponent, update, view

Associated Types

data ComponentState c

The internal state of the component. This is initialized when the component is first shown, and then varies over the lifetime of the component as patchComponent and update modify it.

data ComponentAction c

Internal actions that are emitted by child widgets from the view method and are fed into the update method.

Methods

createComponent :: c event -> (ComponentState c, Maybe (ComponentAction c))

Called when the component is first shown. This creates the initial component state from the declarative component, and optionally provides an initial action.

patchComponent :: ComponentState c -> c event -> ComponentState c

Called when a new declarative component is applied to the existing component state. This method should copy any required state from the new declarative component.

This method will never be called for root components (since the root declarative component never changes) so there is a default implementation that returns an error: components intended to be used as non-root components should implement this method.

update :: c event -> ComponentAction c -> UpdateM c event ()

Called when an event is emitted by a child widget. The UpdateM monad allows updating the component state, running asynchronous IO actions, and sending events to the parent component.

view :: c event -> ComponentState c -> Widget (ComponentAction c)

Provides the tree of widgets that display this component. Called whenever the state changes (and possibly other times too).

data ComponentContext

Keeps track of the current component, and also the global component state, when patching/subscribing-to components.

data UpdateM c event a

An "update action" that is provided by update when a component of type c receives an action from a child widget. The event type parameter describes the type of events that the parent widget can receive.

updateIO :: IO (Maybe (ComponentAction c)) -> UpdateM c event ()

An update action that asynchronously runs some IO - and then maybe emits another action.

updateIO_ :: IO () -> UpdateM c event ()

An update action that asynchronously runs some IO (and never emits another action).

updateParent :: event -> UpdateM c event ()

An update action that (synchronously) sends an action to the parent component.

component :: (Component c, Component parent) => c (ComponentAction parent) -> Widget (ComponentAction parent)

Turn a declarative component into a widget so it can be used as part of a widget tree.

data AppAction x

A top-level component sends this event to its parent when it wants to affect the top-level GTK environment itself.

Constructors

Exit x
Destroy all components, quit GTK, and cause run to finish and produce x.

run :: Component c => c (AppAction x) -> IO x

Start the GTK event loop, then show the component and process its internal events until it emits an `Exit x` event. Then stop GTK and return x to the caller.

The root component should return a GTK Window widget from its view method, otherwise nothing will be visible to the user.

runWith

Arguments

:: Component c

=> IO ()
Additional initialisation action to run after GTK init but before GTK main. E.g. to setup stylesheets.

-> c (AppAction x)

-> IO x

Start the GTK event loop, then show the component and process its internal events until it emits an `Exit x` event. Then stop GTK and return x to the caller.

The root component should return a GTK Window widget from its view method, otherwise nothing will be visible to the user.