GI.Gtk.Declarative.Component.Internal

Documentation

class Typeable c => Component (c :: * -> *) where

A declarative component. Types in this class must have a single type parameter which describes the type of the events emitted by the component.

Minimal complete definition

createComponent, update, view

Associated Types

data ComponentState c

The internal state of the component. This is initialized when the component is first shown, and then varies over the lifetime of the component as patchComponent and update modify it.

data ComponentAction c

Internal actions that are emitted by child widgets and make this component do stuff.

Methods

createComponent :: c event -> (ComponentState c, Maybe (ComponentAction c))

Called when the component is first shown. This creates the initial component state from the declarative component, and optionally provides an initial action to start the component doing stuff.

patchComponent :: ComponentState c -> c event -> ComponentState c

Called when a new declarative component is applied to the existing component state. This method should copy any required state from the new declarative component. This method will never be called for top level components - so there is a default implementation that returns an error: components intended to be used a non-top-level components should implement this method.

update :: c event -> ComponentAction c -> UpdateM c event ()

Called when an action is emitted by a child event. The UpdateM monad allows updating the component state, running asynchronous IO actions, and sending events to the parent on this component.

view :: c event -> ComponentState c -> Widget (ComponentAction c)

Provides the tree of widgets that display this component. Called whenever the state changes (and possibly other times too).

data UpdateM c event a where

An "update action" that happens when a component of type c receives an action from a child widget. The event type parameter describes the type of events that the parent widget can receive. This implements MonadState to allow getting/setting the component state.

Constructors

UpdatePure :: a -> UpdateM c event a

UpdateBind :: UpdateM c event a -> (a -> UpdateM c event b) -> UpdateM c event b

UpdateStateGet :: UpdateM c event (ComponentState c)

UpdateStatePut :: ComponentState c -> UpdateM c event ()

UpdateIO :: IO (Maybe (ComponentAction c)) -> UpdateM c event ()

UpdateParent :: event -> UpdateM c event ()

updateIO :: IO (Maybe (ComponentAction c)) -> UpdateM c event ()

An update action that asynchronously runs some IO - and then maybe emits another action.

updateIO_ :: IO () -> UpdateM c event ()

An update action that asynchronously runs some IO (and never emits another action).

updateParent :: event -> UpdateM c event ()

An update action that (synchronously) sends an action to the parent component.

data UpdateResult x a

The result of running an UpdateM

Constructors

UpdateResultValue

Fields

updateValue :: a

updatedState :: Bool

UpdateResultExited x

updateResult :: UpdateResult x ()

newtype ComponentId

Constructors

ComponentId Int

data StoredComponent

Constructors

forall c parent.(Component c, Component parent) => StoredComponent

Fields

parentComponentId :: !ComponentId

nonRootComponent :: !(c (ComponentAction parent))

storedState :: !(ComponentState c)

storedView :: !(Widget (ComponentAction c))

Must only be updated inside patch, otherwise it gets out-of-sync with the state tree

data ComponentContext

Keeps track of the current component, and also the global component state, when patching/subscribing-to components.

Constructors

ComponentContext

Fields

currentComponentId :: ComponentId

components :: HashMap ComponentId StoredComponent

nextComponents :: IORef (HashMap ComponentId StoredComponent)

nextComponentId :: IORef ComponentId

events :: Chan DynamicAction

data DynamicAction

A dynamically typed action that is destined for a particular component.

Constructors

forall comp.Component comp => DynamicAction ComponentId (ComponentAction comp)