GI.Gtk.Declarative.Container

Description

Implementations for common Gtk.Container.

Documentation

data Container widget children event

Declarative version of a container widget, i.e. a widget with zero or more child widgets. The type of children is parameterized, and differs across the supported container widgets, as some containers require specific types of child widgets. These type relations are decided by IsContainer, and instances can found in GI.Gtk.Declarative.Container.Patch.

container

Arguments

:: (Typeable widget, IsWidget widget, IsContainer widget, FromWidget (Container widget (Children child)) target, ToChildren widget parent child)

=> (ManagedPtr widget -> widget)
A container widget constructor from the underlying gi-gtk library.

-> Vector (Attribute widget event)
Attributes.

-> parent (child event)
The container's child widgets, in a MarkupOf builder.

-> target event
The target, whose type is decided by FromWidget.

Construct a container widget, i.e. a widget with zero or more children.

data Children child event

Common collection type for child widgets, used when patching containers.

class ToChildren widget parent child | widget -> parent, widget -> child where

Converts a specific collection type to Children.

Minimal complete definition

Nothing

Methods

toChildren :: (ManagedPtr widget -> widget) -> parent (child event) -> Children child event

default toChildren :: parent ~ Vector => (ManagedPtr widget -> widget) -> parent (child event) -> Children child event