GI.Gtk.Declarative.Container.Class

Description

Shared interfaces for containers.

Documentation

class IsContainer container child | container -> child where

Describes supported GTK+ containers and their specialized APIs for appending and replacing child widgets.

Methods

appendChild

Arguments

:: container
Container widget

-> child event
Declarative child widget

-> Widget
GTK child widget to append

-> IO ()

Append a child widget to the container.

replaceChild

Arguments

:: container
Container widget

-> Int32
Index to replace at

-> IO ()
IO action to destroy the old child widget

-> child event
New declarative child widget

-> Widget
New GTK widget created from the declarative widget

-> IO ()

Replace the child widget at the given index in the container.

newtype Children child event

Common collection type for child widgets, used when patching containers.

Constructors

Children

Fields

unChildren :: Vector (child event)

class ToChildren widget parent child | widget -> parent, widget -> child where

Converts a specific collection type to Children.

Minimal complete definition

Nothing

Methods

toChildren :: (ManagedPtr widget -> widget) -> parent (child event) -> Children child event

default toChildren :: parent ~ Vector => (ManagedPtr widget -> widget) -> parent (child event) -> Children child event