GI.Gtk.Declarative.Container.Notebook

Description

Implementation of Notebook as a declarative container.

Documentation

data Page event

Describes a page to be added to a Notebook

page :: Text -> Widget event -> Page event

Create a page with a textual label and an arbitrary content widget.

pageWithTab :: Widget event -> Widget event -> Page event

Create a page with arbitrary widgets for both label and content.

notebook :: Vector (Attribute Notebook event) -> Vector (Page event) -> Widget event

Create a Notebook by combining multiple pages.

Orphan instances

IsContainer Notebook Widget

Instance details

Methods

appendChild :: Notebook -> Widget event -> Widget0 -> IO ()

replaceChild :: Notebook -> Int32 -> IO () -> Widget event -> Widget0 -> IO ()

ToChildren Notebook Vector Widget

Instance details

Methods

toChildren :: (ManagedPtr Notebook -> Notebook) -> Vector (Widget event) -> Children Widget event