GI.Gtk.Declarative.Container.Paned

Description

Implementation of Paned as a declarative container.

Documentation

data Pane event

Describes a pane to be packed with panePack1/panePack2 in a Paned.

data PaneProperties

Values used when packing a pane into a Paned.

Constructors

PaneProperties

Fields

resize :: Bool

shrink :: Bool

defaultPaneProperties :: PaneProperties

Defaults for PaneProperties. Use these and override specific fields.

pane :: PaneProperties -> Widget event -> Pane event

Construct a pane to be packed with panePack1/panePack2 in a Paned.

paned :: Vector (Attribute Paned event) -> Pane event -> Pane event -> Widget event

Construct a Paned based on attributes and two child Panes.

Orphan instances

IsContainer Paned Pane

Instance details

Methods

appendChild :: Paned -> Pane event -> Widget -> IO ()

replaceChild :: Paned -> Int32 -> IO () -> Pane event -> Widget -> IO ()