GI.Gtk.Declarative.Container.Patch

Description

This module implements the patch algorithm for containers.

Documentation

class IsContainer container child | container -> child where

Describes supported GTK+ containers and their specialized APIs for appending and replacing child widgets.

Methods

appendChild

Arguments

:: container
Container widget

-> child event
Declarative child widget

-> Widget
GTK child widget to append

-> IO ()

Append a child widget to the container.

replaceChild

Arguments

:: container
Container widget

-> Int32
Index to replace at

-> IO ()
IO action to destroy the old child widget

-> child event
New declarative child widget

-> Widget
New GTK widget created from the declarative widget

-> IO ()

Replace the child widget at the given index in the container.

patchInContainer :: (IsWidget container, IsContainer container, Patchable child, IsContainer container child) => StateTree 'ContainerState container child event cs -> ComponentContext -> container -> Vector (child e1) -> Vector (child e2) -> IO (StateTree 'ContainerState container child event cs)

Patch all children in a container. This does not feature any ID checking, as seen in React, so reordering children in a container can produce many updates.