GI.Gtk.Declarative.EventSource

Description

Event handling for declarative widgets.

Documentation

class EventSource widget where

An EventSource can be subscribed to, with a callback, returning a Subscription.

Methods

subscribe

Arguments

:: ComponentContext

-> widget event
Declarative widget with event handlers.

-> SomeState
State of rendered widget tree.

-> (event -> IO ())
Event callback, invoked on each emitted event until the Subscription is cancelled, or widget is otherwise destroyed.

-> IO Subscription
A Subscription is returned, which can be cancelled.

module GI.Gtk.Declarative.EventSource.Subscription