GI.Gtk.Declarative.EventSource.Subscription

Documentation

data Subscription

A Subscription contains zero or more cancellation actions for connected handlers (to a tree of widgets.) When subscribing to a container widget, all child widgets are also subscribed to, and the Subscriptions are combined using the Semigroup instance.

fromCancellation :: IO () -> Subscription

Create a subscription from a cancellation IO action.

cancel :: Subscription -> IO ()

Cancel a Subscription, meaning that the callback will not be invoked on any subsequent signal emissions.