GI.Gtk.Declarative.Patch

Description

The central patching type class of this library.

Documentation

data Patch

A possible action to take on an existing Widget, decided by the patch method when comparing declarative widgets.

Constructors

Modify (IO SomeState)
An IO action to apply to a Widget to make it reflect an updated declarative widget. The action to apply is calculated from the difference between the old and the new declarative widget.

Replace (IO SomeState)
Replace the current Widget by the widget returned by the IO action.

Keep
Do nothing, i.e. keep the Widget as it is.

class Patchable widget where

A patchable widget is one that can create an underlying GTK widget, or calculate a Patch to be applied to an existing GTK widget that was previously created.

Methods

create :: ComponentContext -> widget e -> IO SomeState

Given a declarative widget that is Patchable, return an IO action that can create a new corresponding Widget. The created widget should be use in corresponding patch modifications, until it is replaced.

patch :: ComponentContext -> SomeState -> widget e1 -> widget e2 -> Patch

Given two declarative widgets of the same widget type (but not necessarily of the same event types,) calculate a Patch.

destroy :: ComponentContext -> SomeState -> widget e -> IO ()

Given a previously created declarative widget, destroy it.