GI.Gtk.Declarative.Widget

Description

A Widget value can wrap any Patchable widget, hiding the underlying widget type, such that you can embed heterogeneous collections of widgets in containers.

Documentation

data Widget event where

A Widget value wraps a Patchable and EventSource widget, providing a constrained equivalent of a Dynamic value. It is used to support heterogeneous containers of widgets, and to support equality checks on different types of widgets when calculating patches.

Constructors

Widget :: (Typeable widget, Patchable widget, EventSource widget) => widget event -> Widget event

Widget to Markup conversion

class FromWidget widget target where

Convert a widget to a target type. This is deliberately unconstrained in it's types, and is used by smart constructors to implement return type polymorphism, so that a smart contructor can return either a Widget, or some specifically typed widget, depending on the context in which it's used.

Methods

fromWidget :: widget event -> target event