GI.Gtk.Declarative.Widget.Conversions

Description

Orphan instances for automatically converting between different widget types. As these use UndecidableInstances, they are kept here in solitary confinement.

Orphan instances

FromWidget widget Widget => FromWidget widget BoxChild
Any widget that can be converted to a Widget can be wrapped as a BoxChild with the default properties.

Instance details

Methods

fromWidget :: widget event -> BoxChild event